refactor: profit을 구하는 메소드와 변수들을 ProfitCalculator로 이동

+++ src/main/java/lotto/util/Converter.java
@@ -5,14 +5,14 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Converter {
-	public static long convertStringOfMoneyToLong(String stringOfMoney) {
+	public static float convertStringOfMoneyToLong(String stringOfMoney) {
 		StringBuilder stringBuilder = new StringBuilder();
 		stringOfMoney.chars().forEach(letter -> {
 			if (letter >= '0' && letter <= '9') {
 				stringBuilder.append((char)letter);
 			}
 		});
-		return Long.parseLong(stringBuilder.toString());
+		return Float.parseFloat(stringBuilder.toString());
 	}
 
 	public static List<Integer> convertStringWithCommaToIntegerList(String string) {
